requirements.txt
streamlit==1.41.1
python-dotenv==1.0.1
langchain==0.3.12
langchain-community==0.3.12
langchain-openai==0.2.12
mysql-connector-python==9.1.0
pypyodbc==1.3.6